Explicitly allow certain file extensions
The MIME type of `.org` files is read as `application/vnd.lotus-organizer`
sometimes. That's the type for files made by Lotus Organizer, a proprietary tool
whose last stable release was in 2003 and which has been discontinued
since 2013.

We're probably not encountering that often, whereas Org-mode files are
relatively plentiful.

This also explicitly permits a few other file extensions that are likely to be
searched, bypassing MIME type checking.