Reorganize interface to take default string query
When I started writing `docsim` I'd thought of it as a tool for comparing
documents (hence `doc`ument `sim`ilarity). But I'd been thinking of "documents"
as "files," not as abstract blobs of text including queries, so I added a
`--query` flag that took a file, not a string.

Previously:

- `--query` took a path argument
- passing in a string query only worked through `STDIN`.

With this commit:

- the first positional argument is always a string query, unless either:
  - there's a Boolean `--stdin` flag which reads the query from `STDIN`, or
  - there's a string flag `--file` that takes a file and acts in the current way.

If there's both a `--file` and `--stdin` flag we error and print a usage message.

Usage lines:

    docsim [OPTION...] QUERY [PATH...]

    command | docsim [OPTION...] --stdin [PATH...]

    docsim [OPTION...] --file PATH [PATH...]

This commit also:

- Updates the README and `man` page accordingly
- Bumps the version to `0.1.4`
- Improves usage error message