Remove --omit-query flag (*always* omit the query)
The query's always going to be 100% similar to itself, so returning it conveys
no useful information, and maintaining this flag complicates tracking scores and
displaying results.

This commit also ensures that the query itself (whether from a positional arg,
`STDIN`, or a `--file` argument) will be parsed and included in the corpus. That
should marginally improve accuracy and also lets us remove the fussy test for
whether the cosine similarity was NaN.