Map terms to int IDs to save memory
Each Document has a termMap, which maps a string term to a float64 weight.
There's a lot of redundant storage there; if a string term appears in a thousand
documents, that string will be stored as a hashed key a thousand times.

Instead, this commit adds a global `termIDs` cache that maps string terms to
integer IDs, with a `termID` function that looks up the term in that cache
and/or creates a new entry.

Primitive benchmarking suggests this reduces peak memory usage by about 15-20%
without increasing processing time. That's not huge, but not bad.

The "benchmarking" used to figure this out:

    $ sudo apt install time
    $ make && /usr/bin/time -v ./docsim "foo bar baz" ~/documents 2>&1 | grep "Maximum resident set size"
    Maximum resident set size (kbytes): 633248