Ensure stemming doesn't panic on certain inputs
Certain odd terms (like "eed") that occasionally appear in code caused the
implementation of the Porter stemming algorithm we'd been using to panic.

The maintainer seems to be AWOL, but someone seems to have forked the original
repo and applied some changes, including fixing that bug.

- old: https://github.com/reiver/go-porterstemmer
- new: https://github.com/blevesearch/go-porterstemmer

This migrates us over to the new version, which has the same interface.